+++ changelog.md
@@ -1,4 +1,7 @@
 Updated to Cobblemon 1.7
 
+## Additions
+- Added compatibility with Cobblemon Size Variations.
+
 ## Changes
 - Fertility is now modifiable via the server config instead of changing the GlobalSpeciesFeature file.